\name{errRate.C}
\alias{errRate.C}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{...}
\description{
 ...
}
\usage{
errRate.C(comb,train,test,threshold)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{comb}{...}
  \item{train}{...}
  \item{test}{...}
  \item{threshold}{...}
}

%\references{ ~put references to the literature/web site here ~ }
\examples{

}
\keyword{distribution}
