\name{errRate}
\alias{errRate}

\title{...}
\description{
 ...
}
\usage{
errRate(x,dataset,threshold)
}

\arguments{
  \item{x}{...}
  \item{dataset}{...}
  \item{threshold}{...}
}

\value{
  \item{}{}
}
%\references{ ~put references to the literature/web site here ~ }
\examples{
 
}
\keyword{distribution}% at least one, from doc/KEYWORDS
