\name{errRate.test}
\alias{errRate.test}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{...}
\description{...
}
\usage{
errRate.test(x,dataset,test,threshold)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{...}
  \item{dataset}{...}
  \item{test}{...}
  \item{threshold}{...}
}

\value{
  ...
}
%\references{ ~put references to the literature/web site here ~ }
\examples{

}
\keyword{distribution}
