\name{mdr.c}
\alias{mdr.c}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{...}
\description{...
}
\usage{
mdr.c(dataset,colresp,cs,combi,cv.fold)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{...}
  \item{colresp}{...}
  \item{cs}{...}
  \item{combi}{...}
  \item{cv.fold}{...}
}

\value{
  ...
}
%\references{ ~put references to the literature/web site here ~ }
\examples{

}
\keyword{distribution}
