\name{ormdr}
\alias{ormdr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{...}
\description{...
}
\usage{
ormdr(dataset,bestcombi,cs,colresp,CI.Asy=TRUE,CI.Boot=FALSE,B=5000)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{...}
  \item{bestcombi}{...}
  \item{cs}{...}
  \item{colresp}{...}
  \item{CI.Asy}{...}
  \item{CI.Boot}{...}
  \item{B}{...}
}

\value{
  ...
}
%\references{ ~put references to the literature/web site here ~ }
\examples{

}
\keyword{distribution}
